requirements.txt
streamlit==1.28.1
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.1
torch==2.1.0
torchvision==0.16.0
tensorflow==2.14.0
mediapipe==0.10.7 